Remove deprecated objects of version 0.2.0
#130

+++ sunshine-junit4/src/test/java/io/github/tatools/sunshine/junit4/JunitStatusTest.java
@@ -16,7 +16,7 @@ public class JunitStatusTest {
     public void codeIfPassed() {
         MatcherAssert.assertThat(
                 new JunitStatus(new FakeResult(true, 0, 0, 0)).code(),
-                Matchers.is(0)
+                Matchers.is((short) 0)
         );
     }
 
@@ -24,7 +24,7 @@ public class JunitStatusTest {
     public void codeIfFailed() {
         MatcherAssert.assertThat(
                 new JunitStatus(new FakeResult(false, 0, 0, 0)).code(),
-                Matchers.is(1)
+                Matchers.is((short) 1)
         );
     }
